+++ examples/standalone-parser/src/types.ts
@@ -4,6 +4,7 @@ export enum OtherEnum {
 }
 
 export const ARR1 = ["A", "B"] as const;
+export type ARR1 = (typeof ARR1)[number];
 
 export const ARR2 = [...ARR1, "C"] as const;