+++ packages/beff-core/src/import_resolver.rs
@@ -116,7 +116,7 @@ impl<'a, R: FsModuleResolver> Visit for ImportsVisitor<'a, R> {
         match &n.decl {
             Decl::TsInterface(n) => {
                 let TsInterfaceDecl { id, .. } = &**n;
-                self.symbol_exports.insert(
+                self.symbol_exports.insert_type(
                     id.sym.clone(),
                     Rc::new(SymbolExport::TsInterfaceDecl {
                         decl: Rc::new(*n.clone()),
@@ -126,7 +126,7 @@ impl<'a, R: FsModuleResolver> Visit for ImportsVisitor<'a, R> {
             }
             Decl::TsEnum(decl) => {
                 let TsEnumDecl { id, .. } = &**decl;
-                self.symbol_exports.insert(
+                self.symbol_exports.insert_type(
                     id.sym.clone(),
                     Rc::new(SymbolExport::TsEnumDecl {
                         decl: Rc::new(*decl.clone()),
@@ -142,7 +142,7 @@ impl<'a, R: FsModuleResolver> Visit for ImportsVisitor<'a, R> {
                     span,
                     ..
                 } = &**a;
-                self.symbol_exports.insert(
+                self.symbol_exports.insert_type(
                     id.sym.clone(),
                     Rc::new(SymbolExport::TsType {
                         ty: Rc::new(*type_ann.clone()),
@@ -162,7 +162,7 @@ impl<'a, R: FsModuleResolver> Visit for ImportsVisitor<'a, R> {
                                 name: name.clone(),
                                 span: it.span,
                             });
-                            self.symbol_exports.insert(name, export);
+                            self.symbol_exports.insert_value(name, export);
                         }
                     }
                 }
@@ -174,46 +174,48 @@ impl<'a, R: FsModuleResolver> Visit for ImportsVisitor<'a, R> {
 
     fn visit_named_export(&mut self, n: &NamedExport) {
         match &n.src {
-            Some(src) => {
+            Some(_src) => {
                 //
                 for s in &n.specifiers {
                     match s {
                         ExportSpecifier::Default(_) => {}
-                        ExportSpecifier::Namespace(ExportNamespaceSpecifier { name, .. }) => {
-                            if let ModuleExportName::Ident(id) = name {
-                                if let Some(file_name) = self.resolve_import(&src.value) {
-                                    self.symbol_exports.insert(
-                                        id.sym.clone(),
-                                        Rc::new(SymbolExport::StarOfOtherFile {
-                                            reference: ImportReference::Star {
-                                                file_name: file_name.clone(),
-                                                span: id.span,
-                                            }
-                                            .into(),
-                                            span: id.span,
-                                        }),
-                                    )
-                                }
-                            }
+                        ExportSpecifier::Namespace(ExportNamespaceSpecifier { .. }) => {
+                            panic!("ExportSpecifier::Namespace(ExportNamespaceSpecifier is not supported")
+                            // if let ModuleExportName::Ident(id) = name {
+                            //     if let Some(file_name) = self.resolve_import(&src.value) {
+                            //         self.symbol_exports.insert(
+                            //             id.sym.clone(),
+                            //             Rc::new(SymbolExport::StarOfOtherFile {
+                            //                 reference: ImportReference::Star {
+                            //                     file_name: file_name.clone(),
+                            //                     span: id.span,
+                            //                 }
+                            //                 .into(),
+                            //                 span: id.span,
+                            //             }),
+                            //         )
+                            //     }
+                            // }
                         }
-                        ExportSpecifier::Named(ExportNamedSpecifier { orig, exported, .. }) => {
-                            if let ModuleExportName::Ident(id) = orig {
-                                let name = match exported {
-                                    Some(ModuleExportName::Ident(ex)) => ex.sym.clone(),
-                                    Some(ModuleExportName::Str(st)) => st.value.clone(),
-                                    None => id.sym.clone(),
-                                };
-                                if let Some(file_name) = self.resolve_import(&src.value) {
-                                    self.symbol_exports.insert(
-                                        name,
-                                        Rc::new(SymbolExport::SomethingOfOtherFile {
-                                            something: id.sym.clone(),
-                                            file: file_name.clone(),
-                                            span: id.span,
-                                        }),
-                                    )
-                                }
-                            }
+                        ExportSpecifier::Named(ExportNamedSpecifier { .. }) => {
+                            panic!("ExportSpecifier::Named(ExportNamedSpecifier is not supported")
+                            // if let ModuleExportName::Ident(id) = orig {
+                            //     let name = match exported {
+                            //         Some(ModuleExportName::Ident(ex)) => ex.sym.clone(),
+                            //         Some(ModuleExportName::Str(st)) => st.value.clone(),
+                            //         None => id.sym.clone(),
+                            //     };
+                            //     if let Some(file_name) = self.resolve_import(&src.value) {
+                            //         self.symbol_exports.insert(
+                            //             name,
+                            //             Rc::new(SymbolExport::SomethingOfOtherFile {
+                            //                 something: id.sym.clone(),
+                            //                 file: file_name.clone(),
+                            //                 span: id.span,
+                            //             }),
+                            //         )
+                            //     }
+                            // }
                         }
                     }
                 }
@@ -302,7 +304,7 @@ pub fn parse_and_bind<R: FsModuleResolver>(
         let renamed = unresolved.renamed;
         let k = (unresolved.name.clone(), unresolved.span);
         if let Some((params, ty)) = locals.content.type_aliases.get(&k) {
-            symbol_exports.insert(
+            symbol_exports.insert_type(
                 renamed.clone(),
                 Rc::new(SymbolExport::TsType {
                     ty: ty.clone(),
@@ -315,7 +317,7 @@ pub fn parse_and_bind<R: FsModuleResolver>(
         }
 
         if let Some(enum_) = locals.content.enums.get(&k) {
-            symbol_exports.insert(
+            symbol_exports.insert_type(
                 renamed,
                 Rc::new(SymbolExport::TsEnumDecl {
                     decl: enum_.clone(),
@@ -326,7 +328,7 @@ pub fn parse_and_bind<R: FsModuleResolver>(
         }
 
         if let Some(intf) = locals.content.interfaces.get(&k) {
-            symbol_exports.insert(
+            symbol_exports.insert_type(
                 renamed,
                 Rc::new(SymbolExport::TsInterfaceDecl {
                     decl: intf.clone(),
@@ -337,29 +339,25 @@ pub fn parse_and_bind<R: FsModuleResolver>(
         }
         if let Some(import) = v.imports.get(&k) {
             match &**import {
-                ImportReference::Named {
-                    orig,
-                    file_name,
-                    span,
-                } => {
-                    let it = Rc::new(SymbolExport::SomethingOfOtherFile {
-                        something: orig.as_ref().clone(),
-                        file: file_name.clone(),
-                        span: *span,
-                    });
-                    symbol_exports.insert(renamed, it);
-                    continue;
+                ImportReference::Named { .. } => {
+                    // let it = Rc::new(SymbolExport::SomethingOfOtherFile {
+                    //     something: orig.as_ref().clone(),
+                    //     file: file_name.clone(),
+                    //     span: *span,
+                    // });
+                    // symbol_exports.insert(renamed, it);
+                    // continue;
+                    panic!("ImportReference::Named is not supported")
                 }
-                ImportReference::Star { span, .. } => {
-                    symbol_exports.insert(
-                        renamed,
-                        Rc::new(SymbolExport::StarOfOtherFile {
-                            reference: import.clone(),
-                            span: *span,
-                        }),
-                    );
-
-                    continue;
+                ImportReference::Star { .. } => {
+                    // symbol_exports.insert(
+                    //     renamed,
+                    //     Rc::new(SymbolExport::StarOfOtherFile {
+                    //         reference: import.clone(),
+                    //         span: *span,
+                    //     }),
+                    // );
+                    panic!("ImportReference::Star is not supported")
                 }
                 ImportReference::Default { .. } => {
                     continue;