+++ packages/beff-core/tests/print_parser.rs
@@ -286,6 +286,18 @@ mod tests {
       "#));
     }
     #[test]
+    fn ok_array_spread2() {
+        insta::assert_snapshot!(ok(r#"
+        export const Arr1 = ["a", "b"] as const
+        export type Arr1 = typeof Arr1[number]
+        export const Arr2 = [...Arr1, "c"] as const
+
+        type Arr2C = typeof Arr2[number];
+
+        parse.buildParsers<{ Arr2C: Arr2C }>();
+      "#));
+    }
+    #[test]
     fn ok_discriminated_union() {
         insta::assert_snapshot!(decoder(
             r#"