+++ packages/beff-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @beff/cli
 
+## 0.0.27
+
+### Patch Changes
+
+- namespace export bug
+
 ## 0.0.26
 
 ### Patch Changes