+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.27
+
+### Patch Changes
+
+- namespace export bug
+- Updated dependencies
+  - @beff/cli@0.0.27
+
 ## 0.0.26
 
 ### Patch Changes